p4i integration Integrate 89669 from 3.1 to main: compile with Borland
[git-p4: depot-paths = "//depot/qt/main/": change = 89670]

+++ src/opengl/qgl.cpp
@@ -2016,7 +2016,7 @@ int QGLWidget::displayListBase( const QFont & fnt, int listBase )
 
 #ifdef Q_WS_WIN // always regenerate disp. lists for pixmaps under Windows
     bool regenerate = glcx->deviceIsPixmap();
-    QString key = fnt.key() + QString::number( regenerate );
+    QString key = fnt.key() + QString::number( (int)regenerate );
 #else
     bool regenerate = FALSE;
     QString key = fnt.key();
